Move helper functions to lib
